Personal Finance Calculator

+++ python_introduction/finance_calculator.py
@@ -1,6 +1,6 @@
 # Personal Finance Calculator
 monthly_income = int(input("Enter your monthly income: "))
-monthly_expense = int(input("Enter your total monthly expenses: "))
+monthly_expense = float(input("Enter your total monthly expenses: "))
 
 monthly_savings = monthly_income - monthly_expense
 print(f"Your monthly savings are {monthly_savings}")
